+++ src/include/test/unit_ivf_flat_group.cc
@@ -182,7 +182,7 @@ TEST_CASE(
   size_t expected_base = 5309;
   size_t expected_partitions = 42;
   size_t expected_temp_size = 314159;
-  size_t expected_dimension = 128;
+  uint32_t expected_dimension = 128;
 
   tiledb::Context ctx;
   tiledb::VFS vfs(ctx);
@@ -371,7 +371,7 @@ TEST_CASE("storage version", "[ivf_flat_group]") {
   size_t expected_base = 9234;
   size_t expected_partitions = 200;
   size_t expected_temp_size = 11;
-  size_t expected_dimension = 19238;
+  uint32_t expected_dimension = 19238;
   auto offset = 2345;
 
   ivf_flat_group x =
